+++ src/main/java/com/buuz135/industrial/proxy/BlockRegistry.java
@@ -64,6 +64,7 @@ public class BlockRegistry {
     public static EnergyFieldProviderBlock energyFieldProviderBlock = new EnergyFieldProviderBlock();
     public static OreDictionaryConverterBlock oreDictionaryConverterBlock = new OreDictionaryConverterBlock();
     public static ProteinReactorBlock proteinReactorBlock = new ProteinReactorBlock();
+    public static ProteinGeneratorBlock proteinGeneratorBlock = new ProteinGeneratorBlock();
 
     public static IFCustomFluidBlock BLOCK_ESSENCE = new IFCustomFluidBlock(FluidsRegistry.ESSENCE, Material.WATER);
     public static IFCustomFluidBlock BLOCK_MILK = new IFCustomFluidBlock(FluidsRegistry.MILK, Material.WATER);