+++ src/main/java/com/buuz135/industrial/tile/block/ProteinGeneratorBlock.java
@@ -3,32 +3,37 @@ package com.buuz135.industrial.tile.block;
 import com.buuz135.industrial.api.book.IPage;
 import com.buuz135.industrial.api.book.page.PageText;
 import com.buuz135.industrial.book.BookCategory;
+import com.buuz135.industrial.proxy.BlockRegistry;
 import com.buuz135.industrial.proxy.ItemRegistry;
-import com.buuz135.industrial.tile.generator.BiofuelGeneratorTile;
+import com.buuz135.industrial.tile.generator.ProteinGeneratorTile;
 import com.buuz135.industrial.utils.RecipeUtils;
-import net.minecraft.block.material.Material;
 import net.minecraft.init.Blocks;
 import net.minecraft.init.Items;
 import net.minecraft.item.ItemStack;
-import net.ndrei.teslacorelib.items.MachineCaseItem;
 
 import java.util.List;
 
-public class BiofuelGeneratorBlock extends CustomOrientedBlock<BiofuelGeneratorTile> {
+public class ProteinGeneratorBlock extends CustomOrientedBlock<ProteinGeneratorTile> {
 
-    public BiofuelGeneratorBlock() {
-        super("biofuel_generator", BiofuelGeneratorTile.class, Material.ROCK, 0, 0);
+
+    public ProteinGeneratorBlock() {
+        super("protein_generator", ProteinGeneratorTile.class);
     }
 
+    @Override
+    public void getMachineConfig() {
+        super.getMachineConfig();
+    }
+
+    @Override
     public void createRecipe() {
         RecipeUtils.addShapedRecipe(new ItemStack(this), "pfp", "imi", "rir",
                 'p', ItemRegistry.plastic,
                 'f', Blocks.FURNACE,
                 'i', Blocks.PISTON,
-                'm', MachineCaseItem.INSTANCE,
-                'r', Items.BLAZE_ROD);
+                'm', BlockRegistry.biofuelGeneratorBlock,
+                'r', Items.SPIDER_EYE);
     }
-
 
     @Override
     public BookCategory getCategory() {
@@ -38,7 +43,7 @@ public class BiofuelGeneratorBlock extends CustomOrientedBlock<BiofuelGeneratorT
     @Override
     public List<IPage> getBookDescriptionPages() {
         List<IPage> pages = super.getBookDescriptionPages();
-        pages.add(0, new PageText("Produces power when provided with Biofuel. A " + PageText.bold("bucket") + " of biofuel will produce 160RF/t, a total " + PageText.bold("640,000") + "RF."));
+        pages.add(0, new PageText("Produces power when provided with Protein. A " + PageText.bold("bucket") + " of biofuel will produce 320RF/t, a total " + PageText.bold("1,280,000") + "RF."));
         return pages;
     }