+++ CHANGELOG.md
@@ -1,8 +1,9 @@
 # 1.5
 + Try to fix [#71](https://github.com/Buuz135/Industrial-Foregoing/issues/71) as it is hard to track
-+ Fixed [#70](https://github.com/Buuz135/Industrial-Foregoing/issues/70
++ Fixed [#70](https://github.com/Buuz135/Industrial-Foregoing/issues/70)
 + Changed `IPlantRecollectable` to `PlantRecollectable` and changed to a IForgeRegistry )
 + Added Protein Reactor with JEI Handler and Craftweaker support.
++ Added Protein Generator that produces 320RF/t with protein.
 
 # 1.4.3
 + Fixed [#65](https://github.com/Buuz135/Industrial-Foregoing/issues/65)